ai-service/requirements.txt
Flask==3.0.0
Pillow==10.0.1
transformers==4.35.0
torch==2.1.0
gunicorn==21.2.0
numpy<2.0.0